Added javadoc for picture file service and changed damage in ram sheep

+++ Main/src/main/java/hills/model/Mob.java
@@ -9,5 +9,5 @@ public abstract class Mob extends Creature{
      */
     protected int damagePoints;
 
-    abstract void dealDamage(double health);
+    abstract void dealDamage(int health);
     }
